requirements.txt
python-telegram-bot==21.6
aiohttp==3.10.10
sqlalchemy==2.0.35
apscheduler==3.10.4
python-dotenv==1.0.1
